+++ qcelemental/tests/test_molecule.py
@@ -212,7 +212,7 @@ def test_water_minima_data(Molecule, water_dimer_minima_data):
     mol_dict["name"] = "water dimer"
     mol = Molecule(orient=True, **mol_dict)
 
-    assert len(mol.pretty_print()) == 661
+    assert len(mol.pretty_print()) == 664
     assert len(mol.to_string("psi4")) == 479
 
     assert sum(x == y for x, y in zip(mol.symbols, ["O", "H", "H", "O", "H", "H"])) == mol.geometry.shape[0]